nvme-rdma: Handle rdma device hot removal
In hot-removal the driver sees a DEVICE_REMOVAL event
for all the active cm_ids (per queue). The driver must
destroy all it's rdma resources by the time all the
DEVICE_REMOVAL events has completed so the device (and
the device driver) can safly remove.

In order to do that, we trigger an implicit controller
deletion on the first event that we see and wait for it
to complete.

This patch fixes multiple use-after-free conditions that
occured due to the device driver unloading before we
freed up all the resources.

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Tested-by: Ming Lin <ming.l@ssi.samsung.com>
Reviewed-by: Jay Freyensee <james.p.freyensee@intel.com>