nvme: Add identify controller fabrics additions
IOCAPSZ for maximum I/O capsule size which we expose
sqe + system page_size, ICDOFF for in-capsule data offset
which we expose zero (not supported) and sglprop which
we don't support sgl in host memory.

In the target, we add a transport specific identify_attrs
callout to fill the relevant data.

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Tested-by: Ming Lin <ming.l@ssi.samsung.com>