nvme/rdma: Use max supported inflight rdma reads
The initiator is the rdma responder, so we set the
maximum supported inflight rdma reads (max_qp_rd_atom).
The target is the rdma requester, so we set the minimum
provided from the peer given max inflight reads and the
device supported max inflight read requests (max_qp_init_rd_atom).

It's rather cryptic I know, but this is what rdma_cm authors
came up with...

Signed-off-by: Steve Wise <swise@chelsio.com>
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>