UBUNTU: snapcraft.yaml: various improvements
BugLink: http://bugs.launchpad.net/bugs/1700577

Update the snapcraft.yaml we ship in Xenial to exploit some of the new feature
present in snapcraft's 3.21.

Among the improvements:

1) include the linux-firmware / raspberry-firmware-wireless package in the final
snap
2) rename it to pi2-kernel to match the snaps we have in the store
3) dynamic versioning
4) autogenerated config from debian.$DEBIAN/config/*
5) automatic download of raspberry-firmware-wireless from ppa:~snappy-dev

This brings the pi2 kernel in line with generic/pc-kernel and
snapdragon/snapdragon-kernel.

The resulting kernel is identical to the one we push to the store, but can be
fully built locally. Tested manually by comparing the two snaps (local generated
and the one we have in the store) and by building an edge image from scratch.

Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>